requirements.txt
Django>=3.0,<4.0
psycopg2>=2.9
python-dotenv>=0.20.0
django-cors-headers==3.11.0
graphene-django==2.15.0
djangorestframework-simplejwt==5.1.0
python-magic==0.4.25
Pillow==9.1.0
requests==2.27.1
certifi==2021.10.8
web3==5.29.1